Main screen - changes

+++ app/src/main/java/com/example/wetter_app/Logic/ImageHandler.kt
@@ -3,7 +3,7 @@ package com.example.wetter_app.Logic
 import com.example.wetter_app.R
 
 class ImageHandler {
-    fun getImageResId(temp: Int, windSpeed: Int): Int {
+    fun getImageResId(temp: Double, windSpeed: Double): Int {
         return when {
             temp > 25 -> R.drawable.sun
             temp < -5 -> R.drawable.ice
